+++ controllers/MailController.php
@@ -219,7 +219,7 @@ class MailController extends Controller
         }
 
         $result = UserPicker::filter([
-            'query' => UserFilter::find()->active()->filterBlockedUsers(),
+            'query' => UserFilter::find()->active()->visible()->filterBlockedUsers(),
             'keyword' => $keyword,
             'permission' => (!Yii::$app->user->isAdmin()) ? new SendMail() : null,
             'disableFillUser' => true,
@@ -309,7 +309,8 @@ class MailController extends Controller
         
         // Preselect user if userGuid is given
         if ($userGuid) {
-            $user = User::findOne(['guid' => $userGuid]);
+            /* @var User $user */
+            $user = User::find()->where(['guid' => $userGuid])->active()->visible()->filterBlockedUsers()->one();
 
             if(!$user) {
                 throw new NotFoundHttpException();